requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
seaborn
matplotlib
tqdm
Pillow